feat(explore): run the new tab's hijacking strip on the Explore hub for anonymous visitors
Search engines land logged-out visitors on the Explore hub — the explore
feed and its sort tabs, the tag and source pages, the directories, the
leaderboard and the Squad directory — and none of those pages made the
new tab's signup pitch. This runs the new tab's hijacking strip there,
on the same `hijacking_variants3` arm, so the surfaces share one
experiment and one design.

The strip moves from the extension into shared as
`auth/HijackingLoginStrip`, with the new-tab-only concerns turned into
props: the extension keeps its layout-v2 gate (the slot it renders in
disappears under v2) and its handoff to the onboarding page (its origin
cannot run social OAuth); the webapp opens auth inline. The headline is
a prop too, because "Own your new tab" is wrong on daily.dev/tags.
`ExploreSignupStrip` is the webapp wrapper: anonymous visitors only,
since members already have the feed it offers.

Two things the move surfaced: the webapp's `onboardingUrl` is a bare
path, so the href is now string-built instead of `new URL()` — which
threw at module load and took every page importing MainFeedLayout down
with it; and the in-feed placement needs `feedGutter` on the strip
itself, because the horizontal inset lives on `<Feed>` rather than on
the page layout.
