+++ packages/shared/src/components/MainFeedLayout.tsx
@@ -24,6 +24,7 @@ import { useFeeds } from '../hooks/feed/useFeeds';
 import { WebappShortcutsRow } from '../features/shortcuts/components/WebappShortcutsRow';
 import { AskSearchBanner } from './marketing/banners/AskSearchBanner';
 import { FeedEngagementBanner } from './brand/FeedEngagementBanner';
+import { ExploreSignupStrip } from './auth/ExploreSignupStrip';
 import FeedContext from '../contexts/FeedContext';
 import feedStyles from './Feed.module.css';
 import AuthContext from '../contexts/AuthContext';
@@ -752,6 +753,11 @@ export default function MainFeedLayout({
   // dropdown, so it's gated on isAnyExplore.
   const showExploreV2PageHeader = isAnyExplore && isV2;
 
+  // The Explore hub's feeds: the sort tabs, the explore-tag feed, and
+  // Discussions — the pages the sidebar's Explore panel leads to.
+  const isExploreFeed =
+    isAnyExplore || isExploreTag || feedName === OtherFeedPage.Discussed;
+
   // v2 also hoists the regular page-header strip up here, OUTSIDE
   // `FeedPageLayoutComponent`, so it can span the full floating-card
   // width without being clamped by `FeedPageLayoutList`'s 680px max
@@ -828,6 +834,9 @@ export default function MainFeedLayout({
             } as CSSProperties
           }
         >
+          {isExploreFeed && (
+            <ExploreSignupStrip className={classNames('mb-4', feedGutter)} />
+          )}
           <FeedEngagementBanner className="mb-3" />
         </div>
         {!isExtension && isHomePage && (
