+++ packages/webapp/pages/sources/index.tsx
@@ -25,6 +25,7 @@ import { ExploreHubHeader } from '@dailydotdev/shared/src/components/header/Expl
 import { PageWrapperLayout } from '@dailydotdev/shared/src/components/layout/PageWrapperLayout';
 import { SourceTopList } from '@dailydotdev/shared/src/components/cards/Leaderboard';
 import { PublicPageSignupBanner } from '@dailydotdev/shared/src/components/auth/PublicPageSignupBanner';
+import { ExploreSignupStrip } from '@dailydotdev/shared/src/components/auth/ExploreSignupStrip';
 import { getLayout } from '../../components/layouts/MainLayout';
 import { getLayout as getFooterNavBarLayout } from '../../components/layouts/FooterNavBarLayout';
 import { defaultOpenGraph } from '../../next-seo';
@@ -132,6 +133,7 @@ const SourcesPage = ({
             }}
           />
         </Head>
+        <ExploreSignupStrip className="mb-6" />
         {!isV2Laptop && (
           <div className="flex justify-between">
             <BreadCrumbs>
